pyproject.toml
[tool.poetry]
name = "telegram-template"
version = "0.1.0"
description = "Простой пример создания бота Telegram используя библиотеку Pyrogram с подключением базы данных через Tortoise-ORM и фреймворком FastAPI для работы с API"
authors = ["Romanov <xxevexxx04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
Pyrogram = "^2.0.102"
TgCrypto = "^1.2.5"
tortoise-orm = "^0.19.3"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
